[#6] Adds missing number to list item
Moves the item to the end of the list, and gives it a new number
so does not renumber everything else. If we were to do this it
might be appropriate to tag that as a release. The point of having the
numbers is to make it easy to reference a checklist item